Round only in the presentation layer

+++ main/utils.py
@@ -3,7 +3,6 @@
 from collections import defaultdict
 from collections.abc import Iterable
 from datetime import date, datetime, timedelta
-from decimal import Decimal
 from typing import Any, cast
 
 from django.contrib.auth import get_user_model
@@ -278,7 +277,7 @@ def get_financial_year_dates() -> tuple[datetime, datetime]:
     return start, end
 
 
-def format_currency(value: Decimal) -> str:
+def format_currency(value: float) -> str:
     """Format a float value as a GBP currency with two decimal places."""
     return f"£{value:.2f}"
